Make Model.Yaml an alias for Model.KeyValue
The key-value model is useful for any configuration file where keys are 
pre-defined and have to be present. Hadoop’s capacity-scheduler.xml
is such a file. If it is empty, YARN will not start. However, if we
use Model.Site to pre-define default values for the different
configuration keys, we loose the helpful descriptions that are stored
in the file.

Actually, Model.Yaml should be similar to Model.Site and use the
key-value pairs stored in the properties HashMap but with different
syntax. The flink-conf.yaml file should be generated with Model.KeyValue
to keep the helpful comments.